part2/requirements.txt
numpy==1.24.1
Pillow==9.3.0
torch==1.13.1+cu117
torchvision==0.14.1+cu117
tqdm==4.64.1
